+++ R/util-aggregate_datasets.R
@@ -1,32 +1,26 @@
-aggregate_datasets <- function(payload, from_IDs = FALSE)
-{
-  if (from_IDs)
-  {
+aggregate_datasets <- function(payload, from_IDs = FALSE) {
+  if (from_IDs) {
+    warn_slow_download(length(payload))
     IDs <- payload
     URLs <- purrr::map_chr(.x = IDs, .f = construct_url)
   }
-  else
-  {
+  else {
+    warn_slow_download(payload$count)
     URLs <- aggregate_result_URLs(payload)
     IDs <- URLs_to_IDs(URLs)
   }
 
-  message("Data download in progress for BacDive-IDs: ", appendLF = FALSE)
+  message(paste0(
+    "Downloading BacDive IDs (",
+    min(as.integer(IDs)),
+    " to ",
+    max(as.integer(IDs)),
+    ", but not necessarily contiguous): "
+  ))
   taxon_data <- purrr::map(URLs, download)
-
-  # Test JSON for contains singly \ escaped characters
-  if (any(grepl("\\n|\\r|\\t", x = taxon_data))) {
-    taxon_data %>%
-    purrr::map(repair_escaping, char = "r") %>%
-    purrr::map(repair_escaping, "n") %>%
-    purrr::map(repair_escaping, "t") ->
-    taxon_data
-  }
-
-  taxon_data <- purrr::map(taxon_data, jsonlite::fromJSON)
   names(taxon_data) <- IDs
 
-  return(taxon_data)
+  taxon_data
 }
 
 
@@ -36,22 +30,18 @@ aggregate_datasets <- function(payload, from_IDs = FALSE)
 #'   `searchTerm` in `retrieve_data()`.
 #'
 #' @return An integer vector of all BacDive URLs within the results.
-aggregate_result_URLs <- function(results)
-{
-  if (length(results$url) == 1)
-    URLs <- results$url
-  else
-  {
-    URLs <- c()
-    while (TRUE) {
-      URLs <- c(URLs, unlist(results$results, use.names = FALSE))
-      if (!is.null(results$`next`))
-        results <- jsonlite::fromJSON(download(results$`next`))
-      else
-        break
+#' @keywords internal
+aggregate_result_URLs <- function(results) {
+  URLs <- list()
+  while (TRUE) {
+    URLs <- append(URLs, unlist(results$results, use.names = FALSE))
+    if (!is.null(results$`next`)) {
+      results <- download(results$`next`)
+    } else {
+      break
     }
   }
-  return(paste0(URLs, "?format=json"))
+  paste0(URLs, "?format=json")
 }