+++ R/util-construct_Renviron_path.R
@@ -2,10 +2,11 @@
 #'
 #' @return A character vector representing the path to the user's `.Renviron`
 #'   file.
-construct_Renviron_path <- function()
-{
+#' @keywords internal
+construct_Renviron_path <- function() {
   path <- Sys.getenv("HOME")
-  if (nzchar(path))
+  if (nzchar(path)) {
     path <- normalizePath("~/")
+  }
   file.path(path, ".Renviron")
 }