+++ R/util-construct_url.R
@@ -1,11 +1,9 @@
 #' Construct URLs of API searches
 #'
 #' @inheritParams retrieve_data
+#' @keywords internal
 construct_url <- function(searchTerm,
-                          searchType = "bacdive_id")
-{
-  searchTerm <- sanitise_input(searchTerm, searchType)
-
+                          searchType = "bacdive_id") {
   utils::URLencode(
     paste0(
       "https://bacdive.dsmz.de/api/bacdive/",