+++ R/util-prepare_Renviron.R
@@ -6,12 +6,11 @@
 #' @export
 #'
 #' @examples \donttest{prepare_Renviron()}
-prepare_Renviron <- function()
-  {
+#' @keywords internal
+prepare_Renviron <- function() {
   r_env_file <- construct_Renviron_path()
 
-  if (!file.exists(r_env_file))
-  {
+  if (!file.exists(r_env_file)) {
     file.create(r_env_file)
     Sys.chmod(r_env_file, mode = "0600")
   }
@@ -26,10 +25,12 @@ prepare_Renviron <- function()
       paste0("^", start_of_line),
       readLines(r_env_file, warn = FALSE)
     ))) {
-      if (type == "email")
+      if (identical(type, "email")) {
         write(paste("\n# Please", message),
-              file = r_env_file,
-              append = TRUE)
+          file = r_env_file,
+          append = TRUE
+        )
+      }
       write(start_of_line, file = r_env_file, append = TRUE)
     }
   }
@@ -43,7 +44,7 @@ prepare_Renviron <- function()
         `file.edit(file.path(Sys.getenv('HOME'), '.Renviron'))`\n
         and", message
       )
-      )
+    )
     file.edit(r_env_file)
   }
 }