+++ tests/testthat/test-construct_url.R
@@ -21,9 +21,4 @@ test_that("URL constructions work", {
     construct_url("Pseudomonas", searchType = "taxon"),
     "https://bacdive.dsmz.de/api/bacdive/taxon/Pseudomonas/?format=json"
   )
-
-  expect_equal(
-    construct_url("Bacillus subtilis spizizenii", searchType = "taxon"),
-    "https://bacdive.dsmz.de/api/bacdive/taxon/Bacillus/subtilis/spizizenii/?format=json"
-  )
 })