+++ tests/testthat/test-sanitise_input.R
@@ -1,0 +1,5 @@
+context("test-sanitise_input")
+
+test_that("Searching with abbr. fails", {
+  expect_error(sanitise_term("B. subtilis"))
+})